# driving/api-rest/contracts/zhuquing-openapi-v3.yaml
openapi: 3.0.1
info:
  title: Pruebat-Demo
  description: Pruebat-Demo Application
  version: v1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /api/v1/trabajadores/{id}:
    get:
      tags:
        - Trabajador api
      summary: Consulta todos los trabajadores
      operationId: getById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResourceResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TrabajadorDto"
    put:
      tags:
        - Trabajador api
      summary: Consulta todos los trabajadores
      operationId: update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrabajadorDto"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResourceResponse"
        "200":
          description: OK
    delete:
      tags:
        - Trabajador api
      summary: Consulta todos los trabajadores
      operationId: delete_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResourceResponse"
        "200":
          description: OK
...
  /api/v1/trabajadores:
    get:
      tags:
        - Trabajador api
      summary: Consulta todos los trabajadores
      operationId: getAll
      parameters:
        - name: pageRequestDto
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/PageRequestDto"
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResourceResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseDtoTrabajadorDto"
    post:
      tags:
        - Trabajador api
      summary: Consulta todos los trabajadores
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrabajadorDto"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResourceResponse"
        "200":
          description: OK
...
components:
  schemas:
    ErrorResource:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        details:
          type: array
          items:
            type: string
    ErrorResourceResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorResource"
...
    AptitudDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        descripcion:
          type: string
    TrabajadorDto:
      required:
        - dni
        - id
        - nombre
        - tiendaId
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        dni:
          type: string
        tiendaId:
          type: integer
          format: int64
        aptitudes:
          type: array
          items:
            $ref: "#/components/schemas/AptitudDto"
...
    Pagination:
      type: object
      properties:
        requestedPage:
          type: integer
          format: int32
        requestedSize:
          type: integer
          format: int32
        retrievedResults:
          type: integer
          format: int32
        totalResults:
          type: integer
          format: int64
        nextPage:
          type: string
        previousPage:
          type: string
    PageRequestDto:
      type: object
      properties:
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        order:
          type: string
    PageResponseDtoTrabajadorDto:
      type: object
      properties:
        pagination:
          $ref: "#/components/schemas/Pagination"
        data:
          type: array
          items:
            $ref: "#/components/schemas/TrabajadorDto"
...
  securitySchemes:
    adfs:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: ""
          scopes: {}